Support for webhook admission API v1
Motivation
----------
v1beta1 and v1 of the webhook admission API are the same, but we need to
recognize them and return the correct API version on the response.

Modifications
-------------
Coerce incoming requests to v1 of the API. Return the same API version
in the response, unless missing in which case we default to v1.