Monitor file system for certificate changes
Motivation
----------
When using mounted Kubernetes secrets, we'd like to update the cert used
for TLS whenever the secret is changed. This will handle things like
cert-manager autorenewing without needing to restart the webhook pods.

Modifications
-------------
Implement a filesystem watcher (based on the nginx-ingress code) to
monitor for changes and apply them to the HTTP server.

Note: fsnotify can't be used directly without the file_watcher
abstraction because fsnotify doesn't handle symlinks well. Symlinks are
important for ConfigMap and Secret mounts. Without this special logic,
only the first change would be recognized and not the second.