Support high security K8S clusters (#20)
Motivation
------------
More secure K8S clusters may require that pods run with
certain security restrictions, such as non-root users.

Modifications
---------------
Change the base image to use a non-root user.

Update the default Shawarma image to one which runs as a
non-root user.

Update the example to apply a securityContext to the webhook
deployment that meets the Restricted policy for K8S.

Update the default sidecar to apply a securityContext to the sidecar
that meets the Restricted policy for K8S.