Use K8S native sidecars
Motivation
----------
Given that Shawarma is a sidecar, it makes the most sense to deploy it
as a native sidecar rather than a regular container.

Modifications
-------------
By default, deploy as a native sidecar. Add an option to disable this
behavior for backward compatibility. This also involved refactoring
sidecar configuration management into a separate file for organizational
purposes.

Using deep copies of the sidecar configuration also reduces the risk of
unexpected behavioral leaks between patch requests.